Use a state enum instead of a simple boolean
The current state of the editing workflow can be saved in the viewmodel.
This allows the edit activity to determine how to act whenever it's
recreated. The state handles showing/hiding the loading symbol, as well
as calling finish() on the edit activity when the cards are done being
saved to the database.